добавил отдельные файлы с данными о жанрах и размерах картинок, добавил проверку на количество жанров
