+++ src/js/backdrop-sizes.js
@@ -1,0 +1,8 @@
+const backdropSizes = {
+    w300: "w300", 
+    w780: "w780",
+    w1280: "w1280", 
+    original: "original"
+  }
+  
+  export default backdropSizes;
