+++ src/js/poster-sizes.js
@@ -1,0 +1,11 @@
+const posterSizes = {
+    w92: "w92",
+    w154: "w154", 
+    w185: "w185", 
+    w342: "w342", 
+    w500: "w500", 
+    w780: "w780", 
+    original: "original"
+  }
+  
+  export default posterSizes;
