+++ src/sass/components/_card-markup.scss
@@ -1,4 +1,4 @@
-#movies {
+#content {
     display: flex;
     flex-wrap: wrap;
 }
